+++ package.json
@@ -6,6 +6,7 @@
     "@babel/preset-react": "^7.18.6",
     "@emotion/react": "^11.10.4",
     "@emotion/styled": "^11.10.4",
+    "@hookform/resolvers": "^2.9.7",
     "@mui/material": "^5.10.3",
     "@react-icons/all-files": "^4.1.0",
     "@reduxjs/toolkit": "^1.8.5",
@@ -20,9 +21,12 @@
     "moment": "^2.29.4",
     "react": "^18.2.0",
     "react-bootstrap": "^2.5.0",
-    "react-calendar": "^3.8.0",
     "react-calendar": "^3.7.0",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.35.0",
+    "scss": "0.2.4",
+    "cssbeautify": "0.3.1",
+    "lodash": "4.17.15",
     "react-icons": "^4.4.0",
     "react-native-fireworks": "^0.2.6",
     "react-redux": "^8.0.2",
