--- src/components/componentsProducts/SelectYear.jsx
@@ -1,38 +1,0 @@
-import * as React from "react";
-import InputLabel from "@mui/material/InputLabel";
-import MenuItem from "@mui/material/MenuItem";
-import FormControl from "@mui/material/FormControl";
-import Select from "@mui/material/Select";
-
-export default function SelectYear() {
-  const [age, setAge] = React.useState("");
-
-  const handleChange = (event) => {
-    setAge(event.target.value);
-  };
-
-  return (
-    <div>
-      <FormControl sx={{ m: 1, minWidth: 120 }}>
-        <InputLabel id="demo-simple-select-helper-label">Exp.year</InputLabel>
-        <Select
-          labelId="demo-simple-select-helper-label"
-          id="demo-simple-select-helper"
-          value={age}
-          label="Exp.month"
-          onChange={handleChange}
-        >
-          <MenuItem value={2022}>2022</MenuItem>
-          <MenuItem value={2023}>2023</MenuItem>
-          <MenuItem value={2024}>2024</MenuItem>
-          <MenuItem value={2025}>2025</MenuItem>
-          <MenuItem value={2026}>2026</MenuItem>
-          <MenuItem value={2027}>2027</MenuItem>
-          <MenuItem value={2028}>2028</MenuItem>
-          <MenuItem value={2029}>2029</MenuItem>
-          <MenuItem value={2030}>2030</MenuItem>
-        </Select>
-      </FormControl>
-    </div>
-  );
-}
