--- src/components/componentsProducts/payBike.css
@@ -1,10 +1,0 @@
-#transition-modal-description {
-  display: flex !important;
-  flex-direction: column !important;
-  flex-wrap: wrap !important;
-  justify-content: flex-start !important;
-  align-items: center !important;
-}
-h2#transition-modal-description {
-  text-align: center;
-}
